requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.40
alembic==1.15.2
psycopg[binary]>=3.2.10
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic-settings==2.9.1
reportlab==4.4.1
python-dotenv==1.1.0
Pillow==11.2.1
aiofiles==24.1.0
pytest==8.3.5
httpx==0.28.1
email-validator==2.1.1
